Cleaned up menu code;
