+++ frontend/pages/agent/[agent].js
@@ -1,8 +1,8 @@
 import { useRouter } from 'next/router';
 import axios from 'axios';
 import useSWR from 'swr';
-import AgentControl from '@/components/agent/AgentControl';
-import ContentSWR from '@/components/content/ContentSWR';
+import AgentControl from '@/components/systems/agent/AgentControl';
+import ContentSWR from '@/components/data/ContentSWR';
 export default function Agent() {
     const agentName = useRouter().query.agent;
     const agent = useSWR(`agent/${agentName}`, async () => (await axios.get(`${process.env.NEXT_PUBLIC_API_URI ?? 'http://localhost:7437'}/api/agent/${agentName}`)).data);
