+++ frontend/pages/chain/index.js
@@ -2,9 +2,9 @@ import axios from 'axios';
 import useSWR from 'swr';
 import ReactMarkdown from 'react-markdown';
 import { Container } from '@mui/material';
-import ContentSWR from '@/components/content/ContentSWR';
-import DoubleSidedMenu from '@/components/content/DoubleSidedMenu';
-import ChainList from '@/components/chain/ChainList';
+import ContentSWR from '@/components/data/ContentSWR';
+import DoubleSidedMenu from '@/components/content/PopoutDrawerWrapper';
+import ChainList from '@/components/systems/chain/ChainList';
 export default function Home() {
   const docs = useSWR('docs/chain', async () => (await axios.get("https://raw.githubusercontent.com/Josh-XT/Agent-LLM/main/docs/concepts/CHAIN.md")).data);
   const chains = useSWR('chain', async () => (await axios.get(`${process.env.NEXT_PUBLIC_API_URI ?? 'http://localhost:7437'}/api/chain`)).data);
