+++ frontend/pages/index.js
@@ -2,7 +2,7 @@ import axios from 'axios';
 import useSWR from 'swr';
 import ReactMarkdown from 'react-markdown';
 import { Container } from '@mui/material';
-import ContentSWR from '@/components/content/ContentSWR';
+import ContentSWR from '@/components/data/ContentSWR';
 export default function Home() {
   const readme = useSWR('docs', async () => (await axios.get("https://raw.githubusercontent.com/Josh-XT/Agent-LLM/main/docs/README.md")).data);
   return <Container>
