+++ frontend/pages/new/agent.js
@@ -4,8 +4,8 @@ import { useState } from 'react';
 import { Container, TextField, Button, Typography } from '@mui/material';
 import { mutate } from "swr"
 import useSWR from 'swr';
-import DoubleSidedMenu from '@/components/content/DoubleSidedMenu';
-import AgentList from '@/components/agent/AgentList';
+import DoubleSidedMenu from '@/components/content/PopoutDrawerWrapper';
+import AgentList from '@/components/systems/agent/AgentList';
 export default function Home() {
   const [name, setName] = useState("");
   const handleCreate = async () => {
