+++ frontend/pages/provider/index.js
@@ -2,9 +2,9 @@ import axios from 'axios';
 import useSWR from 'swr';
 import ReactMarkdown from 'react-markdown';
 import { Container } from '@mui/material';
-import ContentSWR from '@/components/content/ContentSWR';
-import DoubleSidedMenu from '@/components/content/DoubleSidedMenu';
-import ProviderList from '@/components/provider/ProviderList';
+import ContentSWR from '@/components/data/ContentSWR';
+import DoubleSidedMenu from '@/components/content/PopoutDrawerWrapper';
+import ProviderList from '@/components/systems/provider/ProviderList';
 export default function Home() {
   const docs = useSWR('docs/provider', async () => (await axios.get("https://raw.githubusercontent.com/Josh-XT/Agent-LLM/main/docs/concepts/PROVIDER.md")).data);
   const providers = useSWR('provider', async () => (await axios.get(`${process.env.NEXT_PUBLIC_API_URI ?? 'http://localhost:7437'}/api/provider`)).data.providers);
